frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
python-dotenv==1.0.0

